Important bug fixed and framework improved
1. We used to use image.Pixel(i,j,1) to do some processing like feature
description. We should use grayImage.Pixel(i,j,0) instead as the first
one refers to the value in one channel while the second is the gray
scale which is a combination of 3 channels.
2. Rearranged the kd-tree matching speed up method. I treat it as the
4th matching method.
3. Add some instructions for framework and feature/descriptor/matching
types, comments for reference and comprehension.
4. Fixed the text encoding problem for GB2312 which may casue a
termination of the program. Now fit both English system environment and
Chinese.